Enable auto-invite for all CTF channel members
Override 'auto_invite' key to automatically invite everyone in the
channel if the value of the key is set to 'True' instead of to an array
of members.

Any value other than "True" for the key will cause it to function as in
the past, which is to auto invite only the listed members.